Upgrade to Halide 10 with CUDA autoscheduler
Upgrade the Halide toolchain from 8.0 to 10.0, which in turn unlocks
the following features:

 - (completed) to emit GPU accelerated image processing kernels;
 - (completed) to optimize / fuse CUDA kernels with the Li2018 autoscheduler;
 - (completed) supports either x86_64 or ARM environments
   (use-case of Halide on ARM: Nvidia Jetson SoB);
 - (TBD) to derive and optimize the analytic first-order partial derivative
    of any proximal functions, aka the Autodiff;
 - (TBD) to derive and optimize the output-weighted first-order partial
   derivative of any linear operations.

The only catch is that the Halide provided fft routine is not
compatible with Li2018-CUDA autoscheduler. Developers are
encouraged to contribute the cuFFT interfaces to the ProxImaL project.

Likewise, the Affine transform, warp() and warpT(), invokes
the cyclic image boundary conditions; so far the Li2018 autoscheduler
toolchain is unable to optimize such operations in CUDA.
Developers who wish to make use of the feature can contact the
Halide team for help.