Setup Github actions
This change setup the github actions boilerplate in order to test
the proximal project in various (virtual) environments, namely
Python 3.6, 3.7 and 3.8.

This also involves a change in how the Meson build system will search
for the Python C-header files. Previously, Meson looks for the natively
install python runtime on the Ubuntu system. Here, we change the config
to look for the python runtime that is used to run Meson.