ranged for loop -> ranges::transform
Refactor the code to compute the primal and dual residuals, such that
the residual values is captured in the main Halide pipeline.

On clang14+MacOS on the Github CI server, we previously encounter a
problem that the residual stage "r" is captured by r-value, so the
expression for computing the residual was destroyed on the fly. We need
to capture the l-value reference instead. Hence the rewrite.